rest_api/requirements.txt
uvicorn[standard]==0.21.1
fastapi==0.95.0
scylla-driver==3.26.4
requests
pycryptodome==3.17